requirement.txt
streamlit
pymupdf
python-docx
groq
uvicorn
fastapi
requests
python-dotenv
streamlit_mic_recorder
soundfile
streamlit-webrtc
sounddevice
wavio
pyttsx3 
pydub
gTTS
psycopg2-binary 
SQLAlchemy
werkzeug